Update I2CTivaSlave.c
There is a problem where the slave has to write on the I2C bus. If the application never writes on the I2C bus, after the master requests a write, the CLK signal may hang. To solve that, we added the parameter params->writeTimeout, which serves as a guard to write 0x00 in order to avoid crashes. 

The flag object->state.callCallback is also cleared, after a reading, to prevent an accidental post from being generated inside the ISR at the Semaphore readSem. Although the readTaskBlocking() function tries to clear the Semaphore, it is possible that the ISR will slip between Semaphore_pend(Semaphore_handle(&object->readSem), BIOS_NO_WAIT) and RingBuf_get(&object->ringBuffer, &readIn) in line 374.